requirements.txt
websockets==12.0
pandas==2.2.0
numpy==1.26.0
python-telegram-bot==21.0
matplotlib==3.8.0